backend/requirements/base.txt
fastapi
uvicorn
python-dotenv
pydantic
pillow
pytesseract
python-multipart
regex
httpx
openai
azure-ai-formrecognizer
